Add missing `CGFloat` definition
