+++ Sources/Swiftagram/Endpoints/EndpointLocation.swift
@@ -12,28 +12,11 @@ import ComposableRequest
 public extension Endpoint {
     /// A `struct` holding reference to `location` `Endpoint`s. Requires authentication.
     struct Location {
-        /// A `struct` holding reference to longitude and latitude.
-        public struct Coordinates: Equatable {
-            /// The longitude.
-            public var longitude: CGFloat
-            /// The latitude.
-            public var latitude: CGFloat
-
-            /// Init.
-            /// - parameters:
-            ///     - latitude: A `CGFloat` repreenting the latitude.
-            ///     - longitude: A `CGFloat` repreenting the longitude.
-            public init(latitude: CGFloat, longitude: CGFloat) {
-                self.latitude = latitude
-                self.longitude = longitude
-            }
-        }
-
         /// Get locations around coordinates.
         /// - parameters:
         ///     - coordinates: A `CGPoint` with latitude and longitude.
         ///     - query: An optional `String` narrowing down the list. Defaults to `nil`.
-        public static func around(coordinates: Coordinates, matching query: String? = nil) -> Disposable<LocationCollection> {
+        public static func around(coordinates: Swiftagram.Location.Coordinates, matching query: String? = nil) -> Disposable<LocationCollection> {
             return Endpoint.version1
                 .appendingDefaultHeader()
                 .appending(path: "location_search/")
